+++ pkg/server/doc.go
@@ -1,3 +1,2 @@
 // Package server defines the server for dime
-// Package server defines the server for dime
 package server