+++ pkg/server/server.go
@@ -6,9 +6,9 @@ import (
 	"net/http"
 
 	"github.com/gorilla/mux"
+	"github.com/johnmarkli/dime/pkg/store"
 )
 
-// Server manages the lifecycle of the dime
 // Server manages the lifecycle of the dime
 // server
 type Server struct {
@@ -19,13 +19,17 @@ type Server struct {
 func New() (*Server, error) {
 	router := mux.NewRouter()
 
-	hh := healthHandler{}
+	hh := HealthHandler{}
 	router.HandleFunc("/health", hh.Check).Methods("GET")
 
-	dh := dicomHandler{}
+	st := store.NewFileStore("data")
+	dh := NewDICOMHandler(st)
 	dicomsRouter := router.PathPrefix("/dicoms").Subrouter()
-	dicomsRouter.HandleFunc("", dh.UploadDICOM).Methods("POST")
-	dicomsRouter.HandleFunc("", dh.ListDICOMs).Methods("GET")
+	dicomsRouter.HandleFunc("", dh.Upload).Methods("POST")
+	dicomsRouter.HandleFunc("", dh.List).Methods("GET")
+	dicomsRouter.HandleFunc("/{id}", dh.Read).Methods("GET")
+	dicomsRouter.HandleFunc("/{id}/attributes", dh.Attributes).Methods("GET")
+	dicomsRouter.HandleFunc("/{id}/image", dh.Image).Methods("GET")
 
 	s := &Server{
 		server: &http.Server{
@@ -37,17 +41,13 @@ func New() (*Server, error) {
 }
 
 // Run the dime server
-// Run the dime server
 func (s *Server) Run() {
-	slog.Info("Starting dime server")
 	slog.Info("Starting dime server")
 	go s.server.ListenAndServe()
 }
 
 // Shutdown the dime server
-// Shutdown the dime server
 func (s *Server) Shutdown() error {
-	slog.Info("Shutting down dime server")
 	slog.Info("Shutting down dime server")
 	return s.server.Shutdown(context.Background())
 }