+++ pkg/store/doc.go
@@ -1,0 +1,2 @@
+// Package store provides a way to store DICOM images
+package store