+++ CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.0.2]
+
+### Added
+
+- store interface with file store to save DICOM data to filesystem
+- handlers for reading DICOM by SOP Instance UID with routes for reading attributes and getting the image
+- tests for handlers
+
+## [0.0.1]
 
 ### Added